Created OrderEvent/PaymentEvent and SubClasses.
For reuse and common contract purposes I have put certain objects
into common-dtos module so that both order-seervice and payment-service
would use and be up to date, no unintended mismatches can occur.

In short for now OrderEvent, PaymentEvent, OrderStatus, PaymentStatus
have been created. Also PaymentInterface and OrderInterface which will
serve as a contract to keep all the implementing objects bound to a
certain rule set.

I will add more new classes in this module as needed and refactoring on
the way. For now I can say that anything that is used both in order &
payment services should be put here as long as it is not against certain
standards and rules such as OOP, SOLID principles.